Reputation UI Handler and Variable References added
Added a handler to link the Reputation system to the UI, alongside a scene to test that it works.  Handler can be adjusted once we have actual UI assets.

Additionally added Scriptable Object Variables and reference classes to them that can be set to either use one of the scriptable objects or a direct value, allowing for resources used in multiple places to point to the same scriptable object so that less places need to be changed in the value is to be changed for all instances.

Currently the Reputation UI Handler is using the Variable References, but these can be stripped out if it would be preferred to not use them for this project.