Added Data Classes, NotebookHandler, and related scripts
Fixed a few bugs in ScriptableObject variables from previous version.

Added ScripableObjects to hold data about Ailments, Ingredients, and Symptoms.

Added Discoveries class to track discoveries for use in restricting information in the notebook.

Added scripts for a notebook handler that use Data objects to populate various pages.

Notebook includes title page with credits, indices for Ailments, Ingredients, and Symptoms, and entry pages for each of said categories.

Bookmarks allow navigation to the title page and indices as well as 5 bookmarks that can be set to specific entry pages.

Specific entries can also be reached via clicking on the tiles showing their name and icon, be it on an index or another entry's page.

Added tracking of highest reputation tier reached, so entries won't become hidden if you lose reputation.

Added test scene for the Notebook, and test data to populate it.