Updated Notebook, cleaned up some code, and added ingredients, symptoms, and ailments for testing
Notebook now shows entries based on discoveries rather than tier.

Added a chunk of the ingredients, symptoms, and ailments from the Ingredients and Treatments document as best I could based on the information there (Note that as of this update, relations between Symptoms and Ingredients is uncertain, and as such they have been left blank).

Added inspector buttons to indices to automatically sort indices and put contents in the correct tier based on the tier stated in the ScriptableObject.

Added inspector buttons to indices and ailment/ingredient/symptom objects to automatically make related data reciprocate (ie. if clicked for ailments, will make the symptoms and ingredients used in the cures add the ailment in the correct area).

Added inspector buttons to NotebookHandler to automatically add tier 1 ingredients and symptoms to records (discovery list), as well as a separate button to do the same for ailments, and one to automatically discover the symptoms of said ailments.  I would recommend setting the list length of the Discoveries variable of the Record class of the Notebook object to 0 to clear it before using these.  I would also recommend clearing and using these if you edit the various data objects, particularly if you change how many values are in any of their arrays.

Added button to notebook Ailment page to discover the cures for that ailment.

All tier 1 Ingredients, Symptoms, and Ailments are discovered by default, but not cures.

Fixed some various bugs and cleaned up code and removed some duplicate code.  Added some comparison and equivalence functionality that was from VariableRefs and GenericData classes.