Added Icons for Tier 1 & 2 ingredients
Added the icons for Tier 1 & 2 ingredients from the assets drive.

Added VectorGraphics package to use them as svgs.

Updated relevant scripts and prefabs to support  the use of svgs.

Updated the data importer to apply icons to ingredient data objects if available.

Fixed a few typos in icon names so they would be properly detected.