Disabled Icon Settings
Removed some debug script I missed in my prior commit.  As it is no longer needed, it seemed best to not have it cluttering up the console logs.

Also added a variable to set the material and color of the icons for undiscovered ingredients now that there are actual icons to see how it looks.