Added pick up animation
Interacting with a gatherable item will cause the player to start an animation now.  This required interrupting the players movement, and adding two new functions to allow for this, in addition the animation has an attached script to allow PlayerMovement.cs to know when the animation ends and thusly allow it to return control to the player.