Added controls display.
Added displays to show the controls for normal gameplay and to explain how to solve the puzzles.

Moved some editor resources into editor folders so they won't be compiled.

Made test indices for the notebook to display just the ailments/ingredients/symptoms in the test build.

Fixed a few bugs with player movement and the player animation controller.