Support Multiple Indices for Ailment Inflicter
Adjusted the Ailment Inflicter script to support multiple ailment indices being assigned.  This will more easily allow the pool of ailments randomly assigned to NPCs to be expanded in response to specific conditions and story progress.

Updated the credits on the title page of the notebook.

Fixed bug where notebook pages didn't update if you learned something if it was the last page you had open until you went to a different page.

Added note to the assignable bookmarks noting that you can click them to assign them.

Imported the new character models.  However, they are not currently being used as the animations do not work with them at this time.