Immediately inflict random npcs
Added the functions ailNPCs(int count) and ailNPCsTotal(int count) which will immediately select a random [count] npcs from the list to inflict with ailments.  The former will always try to select [count] npcs, regardless of how many are already suffering from ailments, while the latter will try to bring the total number of ailing npcs up to [count] if there are less than that many currently suffering ailments.

Additionally, added a bool to NPCs to indicate if they should begin suffering an ailment again after a delay or not once cured (also impacts if they will automatically suffer an ailment after a delay in the upon the game starting up or when they get added to the ailment inflicter's list.